Only allowing zoom and pan while shift is pressed
