+++ mustard.kv
@@ -184,12 +184,18 @@
                 text: 'Undo'
                 disabled: not root.labeling
                 on_release: root.remove_last_bbox()
+            Button:
+                text: 'Stop'
+                disabled: not root.labeling
+                on_release: root.labeling = False
         ZoomableImage:
             id: scatter
             do_rotation: False
             auto_bring_to_front: False
             scale_min: 0.5
             scale_max: 10
+            clickable_area: [box.x + viewer_buttons.width, box.y, box.width - viewer_buttons.width, box.height]
+            transform_allowed: root.transform_allowed
             on_bbox: root.crop_image(box.x + viewer_buttons.width, box.y, box.width - viewer_buttons.width, box.height)
             Image:
                 id: image
