+++ mustard.py
@@ -53,6 +53,7 @@ from kivy.uix.checkbox import CheckBox
 from kivy.properties import ObjectProperty
 from kivy.properties import StringProperty, NumericProperty
 from kivy.properties import DictProperty
+from kivy.core.window import Window
 
 # To get the graphics, set this as the current working directory
 os.chdir(os.path.dirname(os.path.abspath(__file__)))
@@ -360,6 +361,28 @@ class TimeSlider(Slider):
 class RootWidget(BoxLayout):
     def __init__(self, **kwargs):
         super(RootWidget, self).__init__(**kwargs)
+        self._keyboard = Window.request_keyboard(self._keyboard_closed, self)
+        self._keyboard.bind(on_key_down=self._on_keyboard_down, on_key_up=self._on_keyboard_up)
+
+    def _on_keyboard_down(self, keyboard, keycode, text, modifiers):
+        for viewer in self.data_controller.children:
+            viewer.transform_allowed = 'shift' in modifiers
+            try:
+                viewer.label = int(keycode[1][-1])
+            except ValueError:
+                continue
+            # Return True to accept the key. Otherwise, it will be used by the system.
+        return True
+
+    def _on_keyboard_up(self, keyboard, keycode):
+        for viewer in self.data_controller.children:
+            viewer.transform_allowed = False
+            # Return True to accept the key. Otherwise, it will be used by the system.
+        return True
+
+    def _keyboard_closed(self):
+        self._keyboard.unbind(on_key_down=self._on_keyboard_down)
+        del self._keyboard
 
 
 class Mustard(App):
